backend/requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
Pillow
drf-yasg
requests
celery
rabbitmq
redis
gunicorn
django-cors-headers
django-redis
django-jazzmin
django-filter
django-ckeditor
django-tinymce
django-modeltranslation
django-leaflet
django-geojson
django-map-widgets
django-admin-geomap